Move ConvertService to composetransform package.
(cherry picked from commit 655e48e65309f0191bbbb8b2e9730b6976e148cd)
